# back/docs/openapi3.yaml
swagger: "2.0"
info:
  description: "Une application pour effectuer des relevés médicaux, les analyser et suivre sa santé."
  version: "1.0.0"
  title: "Mon assistance médical"
...
schemes:
- "https"
- "http"
paths:
...
  /utilisateur/{id}/releve:
    get:
      tags:
      - "Utilisateur"
      summary: "Récupération des relevés d'un utilisateur"
      description: "Récupération des relevés d'un utilisateur, possibilité de filtrer par type de relevé et de paginer la récupération par page de 5"
      operationId: "getUserReleve"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "id de l'utilisateur que l'on cherche"
        required: true
        type: "integer"
        format: "int64"
      - name: "page"
        in: "query"
        description: "numero de la page que l'on veut récupérer"
        required: false
        type: "integer"
        format: "int64"
      - name: "type"
        in: "query"
        description: "type des relevés que l'on veut récupérer."
        required: false
        type: "string"
        format: "int64"
      responses:
        "200":
          description: "Relevés trouvés."
          schema:
            $ref: "#/definitions/ListeRelevesMedicaux"
        "500":
          description: "Relevés non trouvés."


########################################
#                                      #
#         Requêtes pour profil         #
#                                      #
########################################
  /profil:
    post:
      tags:
      - "Profil"
      summary: "Création d'un profil utilisateur"
      description: ""
      operationId: "addProfil"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Profil d'utilisateur que l'on veut ajouter à la table."
        required: true
        schema:
          $ref: "#/definitions/Profil"
      responses:
        "200":
          description: "Profil"
          schema:
            $ref: "#/definitions/Profil"
        "500":
          description: "erreur"
    get:
      tags:
      - "Profil"
      summary: "Récupération des profils utilisateur"
      description: ""
      operationId: "getProfils"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        "200":
          description: "Liste des profils"
          schema:
            $ref: "#/definitions/Profils"
        "500":
          description: "erreur"
...
  /releve_medical:
    get:
      tags:
      - "Relevé médical"
      summary: "Récupération des relevés médicaux."
      description: ""
      operationId: "getReleveMedical"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        "200":
          description: "Récupération OK"
          schema:
            $ref: "#/definitions/ListeRelevesMedicaux"
        "500":
          description: "Récupération NOK"
    post:
      tags:
      - "Relevé médical"
      summary: "Création d'un relevé médical"
      description: ""
      operationId: "addReleveMedical"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Relevé medical à ajouter."
        required: true
        schema:
          $ref: "#/definitions/ReleveMedical"
      responses:
        "200":
          schema:
            $ref: "#/definitions/ReleveMedical"
          description: "Création OK"
        "500":
          description: "Création NOK"
    put:
      tags:
      - "Relevé médical"
      summary: "Modification d'un relevé médical"
      description: ""
      operationId: "updateReleveMedical"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Relevé médical à modifier."
        required: true
        schema:
          $ref: "#/definitions/ReleveMedical"
      responses:
        "200":
          description: "Modification OK"
        "500":
          description: "Modification NOK"
  /releve_medical/{id}:
    get:
      tags:
      - "Relevé médical"
      summary: "Récupération des relevés médicaux par son identifiant"
      description: ""
      operationId: "getReleveMedicalById"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifiant du relevé médical."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Récupération OK"
          schema:
            $ref: "#/definitions/ReleveMedical"
        "500":
          description: "Récupération NOK"
    delete:
      tags:
      - "Relevé médical"
      summary: "Suppression d'un relevé médical"
      description: ""
      operationId: "deleteReleveMedicalById"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifiant du relevé médical."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Suppression OK"
        "500":
          description: "Suppression NOK"

########################################
#                                      #
#     Requêtes pour type de relevé     #
#                                      #
########################################
...
  /analyse:
    get:
      tags:
      - "Analyse"
      summary: "Récupération des analyses."
      description: ""
      operationId: "getAnalyse"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        "200":
          description: "Récupération OK"
          schema:
            $ref: "#/definitions/ListeAnalyses"
        "500":
          description: "Récupération NOK"
    post:
      tags:
      - "Analyse"
      summary: "Création d'une analyse."
      description: ""
      operationId: "addAnalyse"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Analyse à ajouter."
        required: true
        schema:
          $ref: "#/definitions/Analyse"
      responses:
        "200":
          description: "Création OK"
          schema:
            $ref: "#/definitions/Analyse"
        "500":
          description: "Création NOK"
    put:
      tags:
      - "Analyse"
      summary: "Modification d'une analyse."
      description: ""
      operationId: "updateAnalyse"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Analyse à modifier."
        required: true
        schema:
          $ref: "#/definitions/Analyse"
      responses:
        "200":
          description: "Modification OK"
        "500":
          description: "Modification NOK"
  /analyse/{id}:
    delete:
      tags:
      - "Analyse"
      summary: "Suppression d'une analyse"
      description: ""
      operationId: "deleteAnalyseById"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifiant de l'analyse à supprimer."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Suppression OK"
        "500":
          description: "Suppression NOK"
  /analyse/relevé/{id}:
    get:
      tags:
      - "Analyse"
      summary: "Récupération des analyses d'un relevé"
      description: ""
      operationId: "getAnalyseByReleve"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifiant du relevé à analyser."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Récupération OK"
          schema:
            $ref: "#/definitions/ListeAnalysesDangerositeCouleur"
        "500":
          description: "Récupération NOK"
          
########################################
#                                      #
#       Requêtes pour dangerosité      #
#                                      #
########################################
  /dangerosite:
    get:
      tags:
      - "Dangerosité"
      summary: "Récupération des dangerosités."
      description: ""
      operationId: "getDangerosite"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        "200":
          description: "Récupération OK"
          schema:
            $ref: "#/definitions/ListeDangerosites"
        "500":
          description: "Récupération NOK"
    post:
      tags:
      - "Dangerosité"
      summary: "Création d'une dangerosité."
      description: ""
      operationId: "addDangerosite"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Dangerosité à ajouter."
        required: true
        schema:
          $ref: "#/definitions/Dangerosite"
      responses:
        "200":
          description: "Création OK"
          schema:
            $ref: "#/definitions/Dangerosite"
        "500":
          description: "Création NOK"
    put:
      tags:
      - "Dangerosité"
      summary: "Modification d'une dangerosité."
      description: ""
      operationId: "updateDangerosite"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Dangerosité à modifier."
        required: true
        schema:
          $ref: "#/definitions/Dangerosite"
      responses:
        "200":
          description: "Modification OK"
        "500":
          description: "Modification NOK"
  /dangerosite/{id}:
    delete:
      tags:
      - "Dangerosité"
      summary: "Suppression d'une dangerosite"
      description: ""
      operationId: "deleteDangerositeById"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifiant de la dangerosite à supprimer."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Suppression OK"
        "500":
          description: "Suppression NOK"

########################################
#                                      #
#         Requêtes pour couleur        #
#                                      #
########################################
  /couleur:
    get:
      tags:
      - "Couleur"
      summary: "Récupération des couleur."
      description: ""
      operationId: "getCouleur"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        "200":
          description: "Récupération OK"
          schema:
            $ref: "#/definitions/ListeCouleurs"
        "500":
          description: "Récupération NOK"
    post:
      tags:
      - "Couleur"
      summary: "Création d'une couleur."
      description: ""
      operationId: "addCouleur"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Couleur à ajouter."
        required: true
        schema:
          $ref: "#/definitions/Couleur"
      responses:
        "200":
          description: "Création OK"
          schema:
            $ref: "#/definitions/Couleur"
        "500":
          description: "Création NOK"
    put:
      tags:
      - "Couleur"
      summary: "Modification d'une couleur."
      description: ""
      operationId: "updateCouleur"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Couleur à modifier."
        required: true
        schema:
          $ref: "#/definitions/Couleur"
      responses:
        "200":
          description: "Modification OK"
        "500":
          description: "Modification NOK"
  /couleur/{id}:
    delete:
      tags:
      - "Couleur"
      summary: "Suppression d'une couleur"
      description: ""
      operationId: "deleteCouleurById"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifiant de la couleur à supprimer."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Suppression OK"
        "500":
          description: "Suppression NOK"
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Profil:
    type: "object"
    required:
    - "label"
    properties:
      id:
        type: "integer"
        format: "int64"
      label:
        type: "string"
        example: "Adulte"
  Profils:
    type: "array"
    items:
      type: "object"
      properties:
        schema:
          $ref: "#/definitions/Profil"
  ReleveMedical:
    type: "object"
    required:
    - "id"
    - "prise_de_mesure"
    - "valeur"
    - "utilisateurId"
    - "typeReleveId"
    properties:
      id:
        type: "integer"
        example: 1
      prise_de_mesure:
        type: "string"
        example: "1997-04-23T00:00:00.511Z"
      valeur:
        type: "integer"
        example: 110
      utilisateurId:
        type: "integer"
        example: 19
      typeReleveId:
        type: "integer"
        example: 6
  ListeRelevesMedicaux:
    type: "array"
    items:
      type: "object"
      properties:
        schema:
          $ref: "#/definitions/ReleveMedical"
  Analyse:
    type: "object"
    required:
    - "id"
    - "mini"
    - "maxi"
    - "dangerositeId"
    - "profilId"
    - "typeReleveId"
    properties:
      id:
        type: "integer"
        example: 12
      mini:
        type: "integer"
        example: 95
      maxi:
        type: "integer"
        example: 110
      dangerositeId:
        type: "integer"
        example: 7
      profilId:
        type: "integer"
        example: 2
      typeReleveId:
        type: "integer"
        example: 3
  ListeAnalyses:
    type: "array"
    items:
      type: "object"
      properties:
        schema:
          $ref: "#/definitions/Analyse"
  Dangerosite:
    type: "object"
    required:
    - "id"
    - "message"
    - "couleurId"
    properties:
      id:
        type: "integer"
        example: 8
      message:
        type: "string"
        example: "A surveiller. Consulter votre médecin."
      couleurId:
        type: "integer"
        example: 2
  ListeDangerosites:
    type: "array"
    items:
      type: "object"
      properties:
        schema:
          $ref: "#/definitions/Dangerosite"
...
  Couleur:
    type: "object"
    required:
    - "id"
    - "label"
    properties:
      id:
        type: "integer"
        example: 1
      label:
        type: "string"
        example: "green"
  ListeCouleurs:
    type: "array"
    items:
      type: "object"
      properties:
        schema:
          $ref: "#/definitions/Couleur"
  DangerositeCouleur:
    type: "object"
    required:
    - "id"
    - "message"
    - "couleur"
    properties:
      id:
        type: "integer"
        example: 8
      message:
        type: "string"
        example: "A surveiller. Consulter votre médecin."
      couleur:
        type: "object"
        properties:
          schema:
            $ref: "#/definitions/Analyse"
  AnalyseDangerositeCouleur:
    type: "object"
    required:
    - "id"
    - "mini"
    - "maxi"
    - "dangerosite"
    - "profilId"
    - "typeReleveId"
    properties:
      id:
        type: "integer"
        example: 12
      mini:
        type: "integer"
        example: 95
      maxi:
        type: "integer"
        example: 110
      dangerosite:
        type: "object"
        properties:
          schema:
            $ref: "#/definitions/DangerositeCouleur"
      profilId:
        type: "integer"
        example: 2
      typeReleveId:
        type: "integer"
        example: 3
  ListeAnalysesDangerositeCouleur:
    type: "array"
    items:
      type: "object"
      properties:
        schema:
          $ref: "#/definitions/AnalyseDangerositeCouleur"